Modified Text Color, Fixed Row & Column Deletion
Previous commits removed the A tag required for row deletion, and
modified the JS arbitrarily [why? It appears a find/replace or similar
broke it] so that it didn't work.

Simple changes fixed it.

Also changed text color of the body of the table because the white with
shadow for outline is unreadable (on my display).